redefine DTO
