+++ src/modules/dto/genre.go
@@ -1,0 +1,6 @@
+package dto
+
+type Genre struct {
+	Genre_id   int    `json:"genreId"`
+	Genre_name string `json:"genreName"`
+}
